--- next/.eslintrc.js
@@ -1,5 +1,0 @@
-const base = require('./.eslintrc.react');
-module.exports = {
-  ...base,
-  extends: ['next', 'next/core-web-vitals', ...base.extends],
-};
