--- next/.eslintrc.react.js
@@ -1,1 +1,0 @@
-../shared-config/.eslintrc.react.js
