+++ shared-config/.eslint.react.mjs
@@ -1,0 +1,5 @@
+const eslintReactConfig = {
+
+};
+
+export default [eslintReactConfig];
