+++ shared-config/.eslintignore
@@ -1,2 +1,3 @@
 build
 vendor
+jest*
