+++ shared-config/OLD.eslintrc.base.js
@@ -1,16 +1,3 @@
-const tsRules = {
-  '@typescript-eslint/ban-ts-comment': 1,
-  '@typescript-eslint/camelcase': 0,
-  '@typescript-eslint/explicit-member-accessibility': 0,
-  '@typescript-eslint/no-empty-function': 0,
-  '@typescript-eslint/no-empty-interface': 0,
-  '@typescript-eslint/no-explicit-any': 1,
-  '@typescript-eslint/no-namespace': 0,
-  '@typescript-eslint/no-non-null-assertion': 1,
-  '@typescript-eslint/no-parameter-properties': 0,
-  '@typescript-eslint/no-unused-vars': 1,
-  '@typescript-eslint/no-var-requires': 1,
-};
 
 const nodeRules = {
   'node/no-extraneous-import': 0,
