+++ shared-config/OLD.eslintrc.react.js
